refactor(middleware): extract durable-grant resolution to a sub-factory
Move grantId/delegationExpiryMs/resolveAgentGrant/resolveExistingGrant/bindGrantOnSuccess into with-kya-os.grants.ts behind createGrantResolution(deps), and introduce the shared MiddlewareDeps bag (with-kya-os.deps.ts). No behaviour change; 1607 tests green. Phase 2a of the >300 split.
